fix(stats): use static export + cache
